Add minijail to whitelist.go
The latest version of crosvm refers to minijail as a separate crate:

https://chromium.googlesource.com/chromiumos/platform/crosvm/+/2dae56768c0dcfb57ff07c051f5785162632073e%5E%21/#F1

The current version of minijail in aosp crosvm-master isn't a crate,
but minijail in aosp master is a crate and builds with crosvm.

This allows the latest version of crosvm to build in the Android tree
and use minijail in the same tree.

Change-Id: I2231149f1488727754430183d04072c5cc0f570a